# okito-server/src/main/resources/docs/openapi.json
{
  "components": {
    "schemas": {
      "PostCommentLikeParam": {
        "properties": {
          "commentId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "809"
          },
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "216"
          },
          "type": {
            "description": "No comments found.",
            "example": "9579"
          }
        }
      },
...
      "PostLikeParam": {
        "properties": {
          "postId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "825"
          },
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "670"
          },
          "type": {
            "description": "No comments found.",
            "example": "12208"
          }
        }
      },
...
      "RespResult": {
        "properties": {
          "timestamp": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "1677208455247"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "107"
          },
          "message": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "No comments found.(object)",
            "type": "object",
            "example": ""
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088/"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/post/comment/like/{userId}": {
      "get": {
        "summary": "select all postCommentLikes by userId",
        "deprecated": false,
        "description": "select all postCommentLikes by userId",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 290,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "13",
            "name": "userId"
          }
        ]
      }
    },
...
    "/post/comment/like": {
      "post": {
        "summary": "add a postCommentLike",
        "deprecated": false,
        "description": "add a postCommentLike",
        "tags": [
          "post comment like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCommentLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"commentId\": 447,\n  \"userId\": 266,\n  \"type\": 30929\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 777,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      },
      "delete": {
        "summary": "delete postCommentLike by id",
        "deprecated": false,
        "description": "delete postCommentLike by id",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 1,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "commentId",
            "required": true,
            "example": "3",
            "name": "commentId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "   userId",
            "required": true,
            "example": "629",
            "name": "userId"
          }
        ]
      },
      "put": {
        "summary": "update a postCommentLike",
        "deprecated": false,
        "description": "update a postCommentLike",
        "tags": [
          "post comment like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCommentLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"commentId\": 735,\n  \"userId\": 399,\n  \"type\": 24941\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 465,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      },
      "get": {
        "summary": "select all postCommentLikes",
        "deprecated": false,
        "description": "select all postCommentLikes",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 340,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlike",
        "parameters": []
      }
    },
...
    "/post/comment/count/post/{postId}": {
      "get": {
        "summary": "count postComments By post id",
        "deprecated": false,
        "description": "count postComments By post id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 740,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentcountpost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "post id",
            "required": true,
            "example": "233",
            "name": "postId"
          }
        ]
      }
    },
...
    "/post/like": {
      "post": {
        "summary": "add a postLike",
        "deprecated": false,
        "description": "add a postLike",
        "tags": [
          "post like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"postId\": 168,\n  \"userId\": 40,\n  \"type\": 25113\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 929,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      },
      "delete": {
        "summary": "delete postLike by id",
        "deprecated": false,
        "description": "delete postLike by id",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 867,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "postId",
            "required": true,
            "example": "921",
            "name": "postId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "query",
            "description": "userId",
            "required": true,
            "example": "650",
            "name": "userId"
          }
        ]
      },
      "put": {
        "summary": "update postLike",
        "deprecated": false,
        "description": "update postLike",
        "tags": [
          "post like controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLikeParam"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"postId\": 407,\n  \"userId\": 115,\n  \"type\": 11588\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 770,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      },
      "get": {
        "summary": "select all postLikes",
        "deprecated": false,
        "description": "select all postLikes",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 202,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike",
        "parameters": []
      }
    },
    "/post/comment": {
      "post": {
        "summary": "add a postComment",
        "deprecated": false,
        "description": "add a postComment",
        "tags": [
          "post comment controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostComment"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 729,\n  \"content\": \"mwycsz\",\n  \"uploadTime\": \"2023-02-24 11:14:15\",\n  \"authorId\": 65,\n  \"postId\": 925,\n  \"parentId\": 829\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 648,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment",
        "parameters": []
      },
      "get": {
        "summary": "select all postComments",
        "deprecated": false,
        "description": "select all postComments",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 151,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment",
        "parameters": []
      }
    },
    "/pin/comment": {
      "post": {
        "summary": "add a pinComment",
        "deprecated": false,
        "description": "add a pinComment",
        "tags": [
          "pin comment controller"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinComment"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 136,\n  \"content\": \"rmqewv\",\n  \"uploadTime\": {},\n  \"authorId\": 392,\n  \"pinId\": 851,\n  \"parentId\": 90\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 913,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment",
        "parameters": []
      },
      "get": {
        "summary": "select all pinComments",
        "deprecated": false,
        "description": "select all pinComments",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 739,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment",
        "parameters": []
      }
    },
...
    "/post/comment/like/page/{userId}": {
      "get": {
        "summary": "select all postCommentLikes by userId and page",
        "deprecated": false,
        "description": "select all postCommentLikes by userId and page",
        "tags": [
          "post comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 275,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "89",
            "name": "userId"
          }
        ]
      }
    },
    "/post/comment/{id}": {
      "delete": {
        "summary": "delete pinComment By id",
        "deprecated": false,
        "description": "delete pinComment By id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 686,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "598",
            "name": "id"
          }
        ]
      },
      "get": {
        "summary": "select postComment By id",
        "deprecated": false,
        "description": "select postComment By id",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 251,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "563",
            "name": "id"
          }
        ]
      }
    },
    "/pin/comment/{id}": {
      "delete": {
        "summary": "delete pinComment By id",
        "deprecated": false,
        "description": "delete pinComment By id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 295,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "878",
            "name": "id"
          }
        ]
      },
      "get": {
        "summary": "select pinComment By id",
        "deprecated": false,
        "description": "select pinComment By id",
        "tags": [
          "pin comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 825,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincomment{id}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "id",
            "required": true,
            "example": "820",
            "name": "id"
          }
        ]
      }
    },
...
    "/post/like/{userId}": {
      "get": {
        "summary": "select all postLikes by userId",
        "deprecated": false,
        "description": "select all postLikes by userId",
        "tags": [
          "post like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 558,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "968",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/comment/like/page/{userId}": {
      "get": {
        "summary": "select all pinCommentLikes by userId and page",
        "deprecated": false,
        "description": "select all pinCommentLikes by userId and page",
        "tags": [
          "pin comment like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 469,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pincommentlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "458",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/like/{userId}": {
      "get": {
        "summary": "select all pinLikes by userId",
        "deprecated": false,
        "description": "select all pinLikes by userId",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 666,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlike{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "userId",
            "required": true,
            "example": "917",
            "name": "userId"
          }
        ]
      }
    },
...
    "/pin/like/page/{userId}": {
      "get": {
        "summary": "select all pinLikes by userId and page",
        "deprecated": false,
        "description": "select all pinLikes by userId and page",
        "tags": [
          "pin like controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 509,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "pinlikepage{userId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  userId",
            "required": true,
            "example": "462",
            "name": "userId"
          }
        ]
      }
    },
...
    "/post/comment/page/post/{postId}": {
      "get": {
        "summary": "select all postComments by postId and page",
        "deprecated": false,
        "description": "select all postComments by postId and page",
        "tags": [
          "post comment controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespResult"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"timestamp\": 1677208455247,\n  \"status\": 258,\n  \"message\": \"success\",\n  \"data\": {}\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "postcommentpagepost{postId}",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "  postId",
            "required": true,
            "example": "296",
            "name": "postId"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "okito",
    "version": "1.0.0"
  },
...
}